refactor(tools): move Args types to public API module
Relocate SearchArgs, CreateArgs, and UpdateArgs from their
command files into tools.rs where they are already re-exported.
Eliminates the re-export indirection and keeps all public API
types in one place. command files now focus purely on logic.

Fixes source_structure test: search.rs was 208 non-test lines
(over the 200-line limit) after SearchArgs was moved into it;
this brings it back to 192.
