+++ src/commands/post.rs
@@ -5,35 +5,7 @@ use crate::client::{CreatePostParams, UpdatePostParams};
 use crate::config::Config;
 
 use crate::output;
-use crate::tools::{SaeError, resolve_client};
-
-#[derive(Debug, clap::Args)]
-pub struct CreateArgs {
-    /// Post title
-    #[arg(long)]
-    pub name: String,
-    /// Post body (Markdown)
-    #[arg(long, conflicts_with = "body_file")]
-    pub body: Option<String>,
-    /// Read body from file (use "-" for stdin)
-    #[arg(long, conflicts_with = "body")]
-    pub body_file: Option<String>,
-    /// Category path
-    #[arg(long)]
-    pub category: Option<String>,
-    /// Tags
-    #[arg(long)]
-    pub tag: Vec<String>,
-    /// Mark as WIP
-    #[arg(long)]
-    pub wip: bool,
-    /// Team name
-    #[arg(long)]
-    pub team: Option<String>,
-    /// Preview without creating (no mutation API calls)
-    #[arg(long)]
-    pub dry_run: bool,
-}
+use crate::tools::{CreateArgs, SaeError, UpdateArgs, resolve_client};
 
 pub(crate) async fn run_get(
     config: &Config,
@@ -72,33 +44,6 @@ pub(crate) async fn run_create(
     };
     let post = client.create_post(team, &params).await?;
     output::action_result("Created", &post, json)
-}
-
-#[derive(Debug, clap::Args)]
-pub struct UpdateArgs {
-    /// Post number
-    pub number: u32,
-    /// New title
-    #[arg(long)]
-    pub name: Option<String>,
-    /// New body (Markdown)
-    #[arg(long, conflicts_with = "body_file")]
-    pub body: Option<String>,
-    /// Read body from file (use "-" for stdin)
-    #[arg(long, conflicts_with = "body")]
-    pub body_file: Option<String>,
-    /// New category path
-    #[arg(long)]
-    pub category: Option<String>,
-    /// New tags (replaces existing)
-    #[arg(long)]
-    pub tag: Vec<String>,
-    /// Team name
-    #[arg(long)]
-    pub team: Option<String>,
-    /// Preview without updating (no mutation API calls)
-    #[arg(long)]
-    pub dry_run: bool,
 }
 
 pub(crate) async fn run_update(
